requirements.txt
pysqlite
psycopg2
simplejson
lxml
pillow
shapely
numpy
django
south
six
tornado
supervisor
